Add shared bit_error_rate utility
