+++ benchmarks/fec_bench.py
@@ -13,6 +13,7 @@ import time
 from pathlib import Path
 
 from genecoder import FEC_REGISTRY, load_plugins
+from genecoder.utils import bit_error_rate
 
 REDUNDANCY_ARGS = {
     "bch": "t",
@@ -24,18 +25,6 @@ REDUNDANCY_ARGS = {
 
 DATA_SIZE = 1_000_000  # 1 MB of random bytes
 ERROR_PROB = 0.01  # probability of flipping each bit
-
-
-def bit_error_rate(original: bytes, recovered: bytes) -> float:
-    """Return the bit error rate between two byte strings."""
-    total_bits = len(original) * 8
-    min_len = min(len(original), len(recovered))
-    errors = 0
-    for o, r in zip(original[:min_len], recovered[:min_len]):
-        errors += (o ^ r).bit_count()
-    if len(recovered) < len(original):
-        errors += (len(original) - len(recovered)) * 8
-    return errors / total_bits if total_bits else 0.0
 
 
 def introduce_bit_errors(data: bytes, prob: float, rng: random.Random | None = None) -> bytes:
